refactor: move chat tab to index
